Migrate zarr admin files
